modules/containerization/code/compose-isolation/backend/requirements.txt
flask
psycopg2-binary  # For connecting to PostgreSQL
